Flytter logikk for flere brevmottakere til sendTilBeslutterSteg
